@@ -10,7 +10,7 @@
     "prefer-stable": true,
     "minimum-stability": "stable",
     "require": {
-        "php": "^7.3"
+        "php": "^8.0"
     },
     "authors": [
         {
@@ -45,7 +45,7 @@
         ]
     },
     "require-dev": {
-        "phpunit/phpunit": "^8.0",
-        "squizlabs/php_codesniffer": "^3.4"
+        "phpunit/phpunit": "^9.5",
+        "squizlabs/php_codesniffer": "^3.5"
     }
 }
